requirements.txt
dash==2.14.2
dash-auth==2.0.0
dash-bootstrap-components==1.5.0
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
gunicorn==21.2.0 

